add a test for replay of messages previous to a restart

+++ fix-gateway-core/src/test/java/uk/co/real_logic/fix_gateway/Timing.java
@@ -25,7 +25,8 @@ import static org.junit.Assert.fail;
 
 public final class Timing
 {
-    public static final long DEFAULT_TIMEOUT_IN_MS = hasDebuggerAttached() ? Long.MAX_VALUE : 5_000;
+    // Long timeout, but one that doesn't cause long overflow.
+    public static final long DEFAULT_TIMEOUT_IN_MS = hasDebuggerAttached() ? Integer.MAX_VALUE : 5_000;
 
     public static <T> T withTimeout(
         final String message,
